Collect all code to install local apps in own dir
