+++ src/launcher/features/apps/appsSlice.ts
@@ -49,7 +49,6 @@ export type State = {
     isConfirmLaunchDialogVisible: boolean;
     confirmLaunchText?: string;
     confirmLaunchApp?: LaunchableApp;
-    isDropZoneVisible: boolean;
 };
 
 const initialState: State = {
@@ -59,7 +58,6 @@ const initialState: State = {
     isLoadingLocalApps: true,
     isLoadingDownloadableApps: true,
     isConfirmLaunchDialogVisible: false,
-    isDropZoneVisible: false,
 };
 
 const equalsSpec = (specOfSoughtApp: AppSpec) => (app: App) =>
@@ -266,14 +264,6 @@ const slice = createSlice({
             state.confirmLaunchText = initialState.confirmLaunchText;
             state.confirmLaunchApp = initialState.confirmLaunchApp;
         },
-
-        // Drop Zone
-        showDropZone(state) {
-            state.isDropZoneVisible = true;
-        },
-        hideDropZone(state) {
-            state.isDropZoneVisible = false;
-        },
     },
 });
 
@@ -285,7 +275,6 @@ export const {
     downloadLatestAppInfoStarted,
     downloadLatestAppInfoSuccess,
     hideConfirmLaunchDialog,
-    hideDropZone,
     installDownloadableAppStarted,
     installDownloadableAppSuccess,
     loadDownloadableAppsError,
@@ -300,7 +289,6 @@ export const {
     setAppIconPath,
     setAppReleaseNote,
     showConfirmLaunchDialog,
-    showDropZone,
     updateAllDownloadableApps,
     updateDownloadableApp,
     updateInstallProgress,
@@ -364,6 +352,3 @@ export const isInProgress = (
     (app.progress.isInstalling ||
         app.progress.isUpgrading ||
         app.progress.isRemoving);
-
-export const isDropZoneVisible = (state: RootState) =>
-    state.apps.isDropZoneVisible;
