+++ src/launcher/features/localAppInstall/dropZoneInfo.module.scss
